chore: consolidate test files, dedupe, and lint the suite
Final cleanup for v2:

Root test files 14 → 9, merged by theme (zero identifier collisions,
no test bodies changed):
- wal_test.go ← wal_test + wal_sim_test + straggler_matrix_test
  (one protocol, one file: state machine, loom-lite sims, matrix)
- property_test.go ← property_test + lifecycle_fuzz_test
  (the lifeModel drives both; they were already cross-dependent)
- runtime_test.go ← runtime_test + compile_config_fuzz_test
- record_test.go ← record_test + dedupe_fuzz_test
- gate_bench_test.go → bench_test.go (consistency)
Module files: typednil_test merged into crash_test in each adapter;
std crash_test merged into middleware_test.

Dedupe: TestCrashHostileRates trimmed to the ±Inf / negative-zero
family — plain out-of-range and NaN rates were already pinned by
TestCompileSentinels.

Lint (whole repo): gofmt clean, gofumpt clean (the repo's formatter
per the Justfile), staticcheck fully clean across all 12 modules.
Fixes while linting: -0.0 literals never produce negative zero
(math.Copysign now does), intentional nil-Context abuse carries
//lint:ignore SA1012 with its charter as the reason.

Validation: root suite + -race, all 12 modules test+vet, fuzz seeds.
