+++ adapter/slog/crash_test.go
@@ -1,10 +1,13 @@
 package slogadapter
 
-// Agent E (adapters leg) — nil and garbage abuse against the bridge.
+// Bridge robustness tests: nil/garbage abuse and typed-nil error
+// containment.
 
 import (
+	"bytes"
 	"context"
 	"log/slog"
+	"os"
 	"testing"
 
 	hc "github.com/happytoolin/happycontext"
@@ -34,4 +37,17 @@ func TestCrashNilAbuse(t *testing.T) {
 	nilSink.Write(context.Background(), rec)
 	New(slog.New(slog.DiscardHandler)).Write(context.Background(), rec)
 	New(slog.Default()).Write(context.Background(), rec)
+}
+
+func TestCrashTypedNilErrorField(t *testing.T) {
+	var pe *os.PathError
+	var buf bytes.Buffer
+	s := &recSink{}
+	rt := hc.MustCompile(hc.Config{Sink: s, SamplingRate: 1})
+	op := hc.Start(context.Background(), rt, hc.OperationStart{Domain: hc.DomainJob, Name: "j"})
+	hc.Add(op.Context(), "e", pe)
+	hc.Error(op.Context(), pe)
+	_ = op.End(nil)
+	rec := s.rec
+	New(slog.New(slog.NewTextHandler(&buf, nil))).Write(context.Background(), rec)
 }
